Rewrite plusMinusCancelsLeft to use `with`-abstraction instead of `rewrite`
The problem with `rewrite` in Idris is that it changes just the goal
type and not the types of any bindings in scope. Hence, in this
particular case the type of the goal is changed from
n + m `minus` n = m
to
m + n `minus` n = m
but the type of `prf` remains ``n `LTE` n + m``, so it does not work
with `plusMinusCancelsRight` anymore as the latter expects
``n `LTE` m + n`` instead, and rewriting the proof type does not work
because the prior rewriting of the goal type also changed the type of
the LTE proof expected by `minus`.

It's good to know some Agda and transfer knowledge between these two
languages!